[refactor](#11) 디렉토리명 변경 및 위치 변경 (learning->progress)
