+++ src/main/java/com/poweranger/hai_duo/progress/api/dto/ChapterResponseDto.java
@@ -1,4 +1,4 @@
-package com.poweranger.hai_duo.quiz.api.dto;
+package com.poweranger.hai_duo.progress.api.dto;
 
 public record ChapterResponseDto(Long chapterId) {
