+++ src/main/java/com/poweranger/hai_duo/progress/domain/entity/Chapter.java
@@ -1,4 +1,4 @@
-package com.poweranger.hai_duo.learning.domain.entity;
+package com.poweranger.hai_duo.progress.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -8,19 +8,11 @@ import java.time.LocalDateTime;
 @Entity
 @NoArgsConstructor
 @Getter
-public class Stage {
+public class Chapter {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long stageId;
-
-    private String stageName;
-
-    @ManyToOne
-    @JoinColumn(name = "chapter_id")
-    private Chapter chapter;
-
-    private int stageNumber; // 챕터 안에서 속하는 스테이지 순서 번호
+    private Long chapterId;
 
     @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
