+++ src/main/java/com/poweranger/hai_duo/progress/domain/repository/LevelRepository.java
@@ -1,6 +1,6 @@
-package com.poweranger.hai_duo.learning.domain.repository;
+package com.poweranger.hai_duo.progress.domain.repository;
 
-import com.poweranger.hai_duo.learning.domain.entity.Level;
+import com.poweranger.hai_duo.progress.domain.entity.Level;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
