+++ src/main/java/com/poweranger/hai_duo/quiz/api/factory/QuizDtoFactory.java
@@ -1,7 +1,7 @@
 package com.poweranger.hai_duo.quiz.api.factory;
 
-import com.poweranger.hai_duo.learning.domain.entity.Chapter;
-import com.poweranger.hai_duo.learning.domain.entity.Stage;
+import com.poweranger.hai_duo.progress.domain.entity.Chapter;
+import com.poweranger.hai_duo.progress.domain.entity.Stage;
 import com.poweranger.hai_duo.quiz.api.dto.*;
 import com.poweranger.hai_duo.quiz.application.reader.QuizReader;
 import com.poweranger.hai_duo.quiz.domain.entity.*;
