+++ src/main/java/com/poweranger/hai_duo/quiz/application/reader/ProgressReader.java
@@ -1,0 +1,10 @@
+package com.poweranger.hai_duo.quiz.application.reader;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class ProgressReader {
+
+}
