+++ src/main/java/com/poweranger/hai_duo/quiz/domain/entity/QuizMeaning.java
@@ -1,6 +1,6 @@
 package com.poweranger.hai_duo.quiz.domain.entity;
 
-import com.poweranger.hai_duo.learning.domain.entity.Stage;
+import com.poweranger.hai_duo.progress.domain.entity.Stage;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
