+++ src/main/java/com/poweranger/hai_duo/quiz/domain/repository/QuizBlankRepository.java
@@ -1,6 +1,6 @@
 package com.poweranger.hai_duo.quiz.domain.repository;
 
-import com.poweranger.hai_duo.learning.domain.entity.Stage;
+import com.poweranger.hai_duo.progress.domain.entity.Stage;
 import com.poweranger.hai_duo.quiz.domain.entity.QuizBlank;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
