+++ src/main/java/com/poweranger/hai_duo/quiz/domain/repository/StageRepository.java
@@ -1,7 +1,7 @@
 package com.poweranger.hai_duo.quiz.domain.repository;
 
-import com.poweranger.hai_duo.learning.domain.entity.Chapter;
-import com.poweranger.hai_duo.learning.domain.entity.Stage;
+import com.poweranger.hai_duo.progress.domain.entity.Chapter;
+import com.poweranger.hai_duo.progress.domain.entity.Stage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
