+++ src/main/java/com/poweranger/hai_duo/user/api/dto/UserDto.java
@@ -2,8 +2,8 @@ package com.poweranger.hai_duo.user.api.dto;
 
 import java.time.LocalDateTime;
 
-import com.poweranger.hai_duo.learning.domain.entity.GameCharacter;
-import com.poweranger.hai_duo.learning.domain.entity.Level;
+import com.poweranger.hai_duo.progress.domain.entity.GameCharacter;
+import com.poweranger.hai_duo.progress.domain.entity.Level;
 import com.poweranger.hai_duo.user.domain.entity.mysql.User;
 
 public record UserDto(
