+++ src/main/java/com/poweranger/hai_duo/user/api/factory/UserFactory.java
@@ -1,7 +1,7 @@
 package com.poweranger.hai_duo.user.api.factory;
 
-import com.poweranger.hai_duo.learning.domain.entity.GameCharacter;
-import com.poweranger.hai_duo.learning.domain.entity.Level;
+import com.poweranger.hai_duo.progress.domain.entity.GameCharacter;
+import com.poweranger.hai_duo.progress.domain.entity.Level;
 import com.poweranger.hai_duo.user.domain.entity.mysql.User;
 import org.springframework.stereotype.Component;
 import java.time.LocalDateTime;
