+++ src/main/java/com/poweranger/hai_duo/user/application/service/UserService.java
@@ -2,17 +2,17 @@ package com.poweranger.hai_duo.user.application.service;
 
 import com.poweranger.hai_duo.global.exception.GeneralException;
 import com.poweranger.hai_duo.global.response.code.ErrorStatus;
-import com.poweranger.hai_duo.learning.api.dto.LevelUpResultDto;
+import com.poweranger.hai_duo.progress.api.dto.LevelUpResultDto;
 import com.poweranger.hai_duo.user.api.dto.UserDto;
-import com.poweranger.hai_duo.learning.api.factory.LevelUpResultFactory;
+import com.poweranger.hai_duo.progress.api.factory.LevelUpResultFactory;
 import com.poweranger.hai_duo.user.api.factory.UserFactory;
-import com.poweranger.hai_duo.learning.application.processor.CharacterUpgradeProcessor;
-import com.poweranger.hai_duo.learning.application.processor.LevelUpProcessor;
-import com.poweranger.hai_duo.learning.domain.entity.GameCharacter;
-import com.poweranger.hai_duo.learning.domain.entity.Level;
+import com.poweranger.hai_duo.progress.application.processor.CharacterUpgradeProcessor;
+import com.poweranger.hai_duo.progress.application.processor.LevelUpProcessor;
+import com.poweranger.hai_duo.progress.domain.entity.GameCharacter;
+import com.poweranger.hai_duo.progress.domain.entity.Level;
 import com.poweranger.hai_duo.user.domain.entity.mysql.User;
-import com.poweranger.hai_duo.learning.domain.repository.CharacterRepository;
-import com.poweranger.hai_duo.learning.domain.repository.LevelRepository;
+import com.poweranger.hai_duo.progress.domain.repository.CharacterRepository;
+import com.poweranger.hai_duo.progress.domain.repository.LevelRepository;
 import com.poweranger.hai_duo.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
